Revert "Feature/DEVS-912"
