--- src/components/Checkbox/styles.module.scss
@@ -1,63 +1,0 @@
-.checkbox {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  position: relative;
-  padding-left: 26px;
-
-  input {
-    position: absolute;
-    opacity: 0;
-    height: 0;
-    width: 0;
-  }
-
-  label {
-    cursor: pointer;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 19px;
-    max-width: 100%;
-    margin-left: 8px;
-
-    @media (max-width: 767px) {
-      font-size: 14px;
-    }
-  }
-}
-
-.checkboxInput {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 24px;
-  height: 24px;
-  border: 1px solid var(--border-color-base);
-  background-color: var(--body-background);
-  cursor: pointer;
-  border-radius: 4px;
-
-  &::after {
-    content: '';
-    position: absolute;
-    display: none;
-    left: 8px;
-    top: 3px;
-    width: 6px;
-    height: 12px;
-    border: solid var(--body-background);
-    border-width: 0 2px 2px 0;
-    -webkit-transform: rotate(45deg);
-    -ms-transform: rotate(45deg);
-    transform: rotate(45deg);
-  }
-}
-
-.checkbox input:checked ~ .checkboxInput {
-  background-color: var(--blue-color);
-  border-color: var(--body-background);
-}
-
-.checkbox input:checked ~ .checkboxInput:after {
-  display: block;
-}
