+++ src/components/WithdrawFees/useWithdrawFees.ts
@@ -1,4 +1,3 @@
-import { signAndSendAllTransactions } from './../../utils/transactions/helpers/signAndSendAllTransactions';
 import { useDispatch } from 'react-redux';
 import { useWallet } from '@solana/wallet-adapter-react';
 
@@ -8,9 +7,11 @@ import { txsLoadingModalActions } from '../../state/txsLoadingModal/actions';
 import { TxsLoadingModalTextStatus } from '../../state/txsLoadingModal/reducers';
 import { notify } from '../../utils';
 import { formatRawSol, NotifyType } from '../../utils/solanaUtils';
+import { createAndSendAllTxns } from '../../utils/transactions';
 import { useConnection } from '../../hooks';
 import { Pair } from '../../state/core/types';
 import { useFetchPair } from '../../requests';
+import { signAndSendAllTransactions } from '../../utils/transactions/helpers/signAndSendAllTransactions';
 
 type UseWithdrawFees = (props: { pool: Pair }) => {
   onWithdrawClick: () => Promise<void>;
