+++ src/pages/CreatePool/components/StepThree.tsx
@@ -1,4 +1,4 @@
-import { FC, useMemo, useState } from 'react';
+import { FC, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { PairType } from 'hadeswap-sdk/lib/hadeswap-core/types';
@@ -21,7 +21,6 @@ import { getRawDelta, getRawSpotPrice } from '../../../utils';
 
 import styles from './styles.module.scss';
 import usePriceGraph from '../../../components/Chart/hooks/usePriceGraph';
-import TransactionsWarning from '../../../components/TransactionsWarning/TransactionsWarning';
 
 interface StepThreeProps {
   pairType: PairType;
@@ -85,9 +84,6 @@ export const StepThree: FC<StepThreeProps> = ({
 
   const creationSpotPrice = Math.ceil(formValue.spotPrice * 1e9);
 
-  const [isSupportSignAllTxns, setIsSupportSignAllTxns] =
-    useState<boolean>(true);
-
   const { create: onCreatePoolClick } = useCreatePool({
     pairType,
     buyOrdersAmount,
@@ -98,7 +94,6 @@ export const StepThree: FC<StepThreeProps> = ({
     rawDelta: rawDelta,
     rawFee,
     onAfterTxn: () => history.push('/my-pools'),
-    isSupportSignAllTxns,
   });
 
   const chartData = usePriceGraph({
@@ -121,10 +116,6 @@ export const StepThree: FC<StepThreeProps> = ({
         <Spinner />
       ) : (
         <>
-          <TransactionsWarning
-            checked={!isSupportSignAllTxns}
-            onChange={() => setIsSupportSignAllTxns(!isSupportSignAllTxns)}
-          />
           <div className={styles.settingsBlock}>
             <PriceBlock
               ref={priceBlockRef}
