+++ src/pages/MyPools/hooks/useWithdrawAllFees.ts
@@ -1,4 +1,3 @@
-import { signAndSendAllTransactions } from './../../../utils/transactions/helpers/signAndSendAllTransactions';
 import { useDispatch } from 'react-redux';
 import { useWallet } from '@solana/wallet-adapter-react';
 import {
@@ -7,11 +6,13 @@ import {
 } from '../../../components/TransactionsLoadingModal';
 import { txsLoadingModalActions } from '../../../state/txsLoadingModal/actions';
 import { TxsLoadingModalTextStatus } from '../../../state/txsLoadingModal/reducers';
+import { createAndSendAllTxns } from '../../../utils/transactions';
 import { useConnection } from '../../../hooks';
 import { createWithdrawLiquidityAllFeesTxns } from '../../../utils/transactions/createWithdrawLiquidityAllFeesTxns';
 import { notify } from '../../../utils';
 import { NotifyType } from '../../../utils/solanaUtils';
 import { Pair } from '../../../state/core/types';
+import { signAndSendAllTransactions } from '../../../utils/transactions/helpers/signAndSendAllTransactions';
 
 type UseWithdrawAllFees = (props: { pairs: Pair[] }) => {
   onWithdrawClick: () => Promise<void>;
