+++ src/utils/transactions/helpers/index.ts
@@ -1,6 +1,7 @@
-export * from './signAndSendTransaction';
-export * from './signAndSendAllTransactions';
 export * from './signAndSendTransactionsInSeries';
 export * from './signAndSendAllTransactionsInSeries';
 export * from './showSolanaFMLinkNotification';
+export * from './signAndConfirmTransaction';
+export * from './createAndSendAllTxns';
+export * from './createAndSendTxn';
 export * from './getArrayByNumber';
